requirements.txt
pillow
piexif
fastapi
uvicorn
python-multipart